Squashed commit of the following:
commit f5f55188775af698370630056aae3050d1ed2f98
Merge: e7ceb2e09 5eb7f4c95
Author: Andi Fahruddin Akas <andifahruddinakas@gmail.com>
Date:   Sun Oct 31 14:18:58 2021 +0800

    Merge commit '5eb7f4c95f385616d2e55121c84f8078bcd82823' into testing-2

commit 5eb7f4c95f385616d2e55121c84f8078bcd82823
Author: Andi Fahruddin Akas <andifahruddinakas@gmail.com>
Date:   Sun Oct 31 14:18:49 2021 +0800

    Squashed 'themes/natra/' changes from d8a90d631..41a3c3fbc

    41a3c3fbc Jangan tampilkan data analisis jika tdk tersedia.
    c673b7e37 Cleaner dan Minifikasi  (style.css dan theme.css)

    git-subtree-dir: themes/natra
    git-subtree-split: 41a3c3fbcf68b510ab4add631ce58a84aed205e2

commit e7ceb2e090aae760c1686dc241cb0bd00edceda1
Author: Andi Fahruddin Akas <andifahruddinakas@gmail.com>
Date:   Sun Oct 31 11:16:29 2021 +0800

    Hasil testing

    Perbaikan:

    1. Tambahkan referensi pekerjaan pada indikator.
    2. Perbaiki mode normal/fullscreen pada input data surver yg tdk jalan.
    3. Default value bobot indikator dan pertanyaan.
    4. Perbaiki tipe referensi  untuk tgl lahir pd analisis indikator.
    5. Perbaiki update indikator tdk menghapus jawaban jika menggunakan referensi.
    6. Jangan tampilkan data analisis jika tdk tersedia.

+++ donjo-app/models/Analisis_indikator_model.php
@@ -126,20 +126,29 @@
 
 	private function validasi_data($post)
 	{
-		$data = array();
-		$data['id_tipe'] = $post['id_tipe'] ?: null;
-		$data['nomor'] = nomor_surat_keputusan($post['nomor']);
-		$data['pertanyaan'] = htmlentities($post['pertanyaan']);
-		$data['id_kategori'] = $post['id_kategori'] ?: null;
-		$data['bobot'] = bilangan($post['bobot']);
-		$data['act_analisis'] = $post['act_analisis'];
-		$data['is_publik'] = $post['is_publik'];
-		$data['referensi'] = $post['referensi'];
+		$data = [
+			'id_tipe' => $post['id_tipe'],
+			'referensi' => $post['referensi'] ?? null,
+			'nomor' => nomor_surat_keputusan($post['nomor']),
+			'pertanyaan' => htmlentities($post['pertanyaan']),
+			'id_kategori' => $post['id_kategori'] ?? null,
+			'bobot' => bilangan($post['bobot']),
+			'act_analisis' => $post['act_analisis'],
+			'is_publik' => $post['is_publik'],
+		];
+
+		if (! $post['referensi'])
+		{
+			unset($data['id_tipe']);
+			unset($data['referensi']);
+		}
+
 		if ($data['id_tipe'] != 1)
-			{
-				$data['act_analisis'] = 2;
-				$data['bobot'] = 0;
-			}
+		{
+			$data['act_analisis'] = 2;
+			$data['bobot'] = 0;
+		}
+
 		return $data;
 	}
 
@@ -191,7 +200,7 @@
 
 		$data = $this->validasi_data($this->input->post());
 
-		if ($data['id_tipe'] == 1 OR $data['id_tipe'] == 2)
+		if (($data['id_tipe'] == 1 OR $data['id_tipe'] == 2) && ! $data['referensi'])
 		{
 			$this->db->where('id_indikator', $id)->delete('analisis_parameter');
 		}
@@ -406,7 +415,6 @@
 					],
 					'tanggallahir' => [
 						'judul' => 'Tanggal Lahir',
-						'tipe' => 5,
 					],
 					'agama_id' => [
 						'judul' => 'Agama',
@@ -422,6 +430,11 @@
 						'judul' => 'Pendidikan Sedang Ditempuh',
 						'tipe' => 1,
 						'referensi' =>  $this->referensi_model->list_data('tweb_penduduk_pendidikan'),
+					],
+					'pekerjaan_id' => [
+						'judul' => 'Pekerjaan',
+						'tipe' => 1,
+						'referensi' =>  $this->referensi_model->list_data('tweb_penduduk_pekerjaan'),
 					],
 					'status_kawin' => [
 						'judul' => 'Status_perkawinan',
@@ -458,7 +471,7 @@
 					],
 					// id_cluster => wilayah, agar tdk duplikasi
 					'wilayah' => [
-						'judul' => 'Wilayah (Dusun/RW/T)',
+						'judul' => 'Wilayah (Dusun/RW/RT)',
 					],
 					'status' => [
 						'judul' => 'Status Penduduk',
@@ -590,9 +603,6 @@
 			// Keluarga
 			default:
 				$data = [
-					'no_kk' => [
-						'judul' => 'No KK',
-					],
 					'nik_kepala' => [
 						'judul' => 'NIK Kepala KK',
 					],
@@ -606,7 +616,7 @@
 					],
 					// id_cluster => wilayah, agar tdk duplikasi
 					'wilayah' => [
-						'judul' => 'Wilayah (Dusun/RW/T)',
+						'judul' => 'Wilayah (Dusun/RW/RT)',
 					],
 				];
 				break;
